refactor(#84): Icon 컴포넌트 리팩토링
